Notifications for waiting Processes and auto release
Waiting Processes receive notifications now when the Locks they are
waiting for become available to wake them. Moreover the Locks held by a
Process are free'd now when it is destroyed.

This is something like a first alfa Release that could be used, it seems
complete and working. Of course, some features (like destroying Locks,
dealing with messages / Connections being lost or querying a Lock) are
not implmented yet, but I think these features are not adding
essential power in a theoretical since (what `can be done' with the lock
manager, not power like computational power) but just convenience.
Hence, all most crucial features may be implemented.