Multiple locks by same process on one level
Now, we use multisets to store the `I' and S locks in order to allow for
mulitple locks on one level held by the same process (this was an
embarassing mistake ...). Moreover only one of potentially many lock
requests is removed on release (not used yet but the same kind of bug).